django_jwt/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
